% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{coef.h2ogpumlnet}
\alias{coef.h2ogpumlnet}
\title{h2ogpumlnet coefficients.}
\usage{
\method{coef}{h2ogpumlnet}(object, s = NULL, ...)
}
\arguments{
\item{object}{Value returned by h2ogpumlnet(..).}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Returns coeficients of h2ogpumlnet fit.
}

