% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{cv.h2ogpumlnet}
\alias{cv.h2ogpumlnet}
\title{Cross-validation for glmnet}
\usage{
cv.h2ogpumlnet(x, y, weights, lambda = NULL, nfolds = 10, foldid, ...)
}
\arguments{
\item{x}{Data matrix.}

\item{y}{Response vector}

\item{weights}{Weights of each training example, defaults to 1.}

\item{lambda}{List of lambda values for which CV should be performed.}

\item{nfolds}{Number of folds.}

\item{foldid}{Optional vector of values between 1 and nfolds, assigning
datapoint to cv-fold.}

\item{...}{Other parameters to h2ogpumlnet.}
}
\description{
Does k-fold cross-validation for glmnet.
}

