% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{kLogistic}
\alias{kLogistic}
\title{Logistic Function}
\usage{
kLogistic(m = 1)
}
\arguments{
\item{m}{Number of times that the function should be repeated.}
}
\description{
\deqn{h(x) = log(1 + e^x)}
}

