% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{kMaxPos0}
\alias{kMaxPos0}
\title{Positive-Part Function}
\usage{
kMaxPos0(m = 1)
}
\arguments{
\item{m}{Number of times that the function should be repeated.}
}
\description{
\deqn{h(x) = max(0, x)}
}

