% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{plot.cv.h2ogpumlnet}
\alias{plot.cv.h2ogpumlnet}
\title{Plot h2ogpumlnet CV}
\usage{
\method{plot}{cv.h2ogpumlnet}(x, sign.lambda, ...)
}
\arguments{
\item{x}{h2ogpumlnet.cv object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative if sign.lambda=-1.}

\item{...}{Other graphical parameters to plot.}
}
\description{
Plots results of h2ogpumlnet CV
}

