% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{plot.h2ogpumlnet}
\alias{plot.h2ogpumlnet}
\title{h2ogpumlnet coefficient plot.}
\usage{
\method{plot}{h2ogpumlnet}(x, xvar = c("norm", "lambda", "dev"), ...)
}
\arguments{
\item{x}{h2ogpumlnet object.}

\item{xvar}{x-variable to plot coefficient against.}

\item{...}{Extra parameters for matplot.}
}
\description{
Plots coefficients
}

