% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{predict.h2ogpumlnet}
\alias{predict.h2ogpumlnet}
\title{H2OGPUMLnet prediction}
\usage{
\method{predict}{h2ogpumlnet}(object, newx, s = NULL, type = c("link",
  "response", "class"), ...)
}
\arguments{
\item{object}{Value returned by h2ogpumlnet(..).}

\item{newx}{New value of x for which prediction should be done.}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{type}{Predict response or class.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Predict response for new value input
}

