% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2ogpuml.R
\name{print.h2ogpumlnet}
\alias{print.h2ogpumlnet}
\title{Prints h2ogpumlnet summary}
\usage{
\method{print}{h2ogpumlnet}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{h2ogpumlnet object}

\item{digits}{Number of digits to display}

\item{...}{Not used.}
}
\description{
Prings h2ogpumlnet summary
}

